Move String extensions to a common file
- Move `components()` from NumAsHexProtocol.swift
- Move `splitToShortStrings()` from ByteArray.swift
