+++ Sources/Starknet/Data/Numbers/NumAsHexProtocol.swift
@@ -110,16 +110,6 @@ public extension NumAsHexProtocol {
     }
 }
 
-extension String {
-    func components(withMaxLength length: Int) -> [String] {
-        stride(from: 0, to: self.count, by: length).map {
-            let start = self.index(self.startIndex, offsetBy: $0)
-            let end = self.index(start, offsetBy: length, limitedBy: self.endIndex) ?? self.endIndex
-            return String(self[start ..< end])
-        }
-    }
-}
-
 public enum HexStringDecodingError: Error {
     case invalidStringFormat
 }
